correctly refactor pipes_demo to match new data driven approach
