+++ demo/DTU.py
@@ -1,3 +1,7 @@
+import sys
+import os
+sys.path.append(os.path.join(os.path.dirname(__file__), '../utils'))  # Dynamically add the utils directory to the path
+
 from scene import Scene
 import taichi as ti
 from taichi.math import *
@@ -9,9 +13,6 @@ scene.set_floor(0, (1.0, 1.0, 1.0)) # Height of the floor
 scene.set_background_color((0.5, 0.5, 0.4)) # Color of the sky
 scene.set_directional_light((1, 1, -1), 0.1, (1, 0.8, 0.6)) # Direction and color of the light
 
-
-
-
 # Define the pattern
 pattern = [
     "##########RRR####R###RRRR#####",
@@ -21,8 +22,6 @@ pattern = [
     "##########RRR##RRRRR#R##R#####"
 ]
 
-
-
 @ti.kernel
 def initialize_voxels():
     # Loop through the pattern and set red voxels
@@ -30,10 +29,6 @@ def initialize_voxels():
         for x in ti.static(range(len(pattern[y]))):
             if pattern[y][x] == 'R':
                 scene.set_voxel(ivec3(x, y, 0), 2, vec3(0.8, 0.0, 0.0))  # Corporate red voxel
-
-
-
-
 
 initialize_voxels()
